Move main method out of BoundingBoxMamut
